% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cat_or_zcat}
\alias{cat_or_zcat}
\title{select zstdcat, zcat, or cat}
\usage{
cat_or_zcat(
  filename,
  configs = list(zstdcat.path = "zstdcat", zcat.path = "zcat")
)
}
\arguments{
\item{f}{A file path.}
}
\value{
One of the following 3 commands: zstdcat, zcat, or cat.
}
\description{
Check the suffix of the filename and return one of zstdcat, zcat, or cat
}
\examples{
cat_or_zcat("file.txt")
cat_or_zcat("file.txt.gz")
cat_or_zcat("file.txt.zst")
\dontrun{
cat_or_zcat(file_path)
fread(cmd = paste(cat_or_zcat(file_path), file_path))
}

}
