% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{compose_regression_formula_str}
\alias{compose_regression_formula_str}
\title{Compose a regression formula give a response variable and the predictors}
\usage{
compose_regression_formula_str(response, predictors, quote_char = "`")
}
\arguments{
\item{response}{The response variable}

\item{predictors}{The predictor variables}

\item{quote_char}{A character for quotation of terms}
}
\value{
A string representing the regressino formula
}
\description{
response ~ 1 + predictor[1] + predictor[2] + ... + predictor[n]
}
\details{
The function quotes variable names using quote_char
}
\examples{
compose_regression_formula_str('HC326', c('covar', 'PRS.score-1'))
compose_regression_formula_str('HC326', c('covar', 'PRS.score-1'), quote_char='')

}
