% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_matrix_product}
\alias{compute_matrix_product}
\title{Take a matrix product of data matrix (X_df) and coefficient matrix (beta_df)}
\usage{
compute_matrix_product(
  X_df,
  beta_df,
  variables = NULL,
  beta_estimate_cols = c("estimate"),
  beta_variable_col = "variable"
)
}
\arguments{
\item{X_df}{A data frame containing the dataset (X)}

\item{beta_df}{A string representing a regression formula}

\item{variables}{A subset of variables you'd like to take the product on}

\item{beta_estimate_cols}{The set of column names of the beta_df that contains the values of the coefficients.}

\item{beta_variable_col}{The column name of the beta_df that contains the variable name.}
}
\value{
a data frame containing the results of XB
}
\description{
Given a data matrix (X_df) and a coefficient matrix (beta_df),
compute the matrix product.
You can specify the set of variables (variables) to focus
if you don't want to use all the data in the beta_df.
}
\examples{
\dontrun{
compute_matrix_product(data_df, glm_fit_df, c('age', 'sex'), c('estimate'), 'variable')
}

}
