% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{config_params_data_type}
\alias{config_params_data_type}
\title{This is a helper function for read_config_from_file().
Here, we define the default data types so that we can properly parse
the parameters specified in the config file.}
\usage{
config_params_data_type()
}
\description{
This is a helper function for read_config_from_file().
Here, we define the default data types so that we can properly parse
the parameters specified in the config file.
}
