% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count_n_per_split}
\alias{count_n_per_split}
\title{Count the number of individuals in each popluation split}
\usage{
count_n_per_split(df, pheno_col, family, case_code = 1)
}
\arguments{
\item{df}{A phenotype data frame with a column named 'split'}
}
\value{
a data frame containing the number of individuals
}
\description{
Count the number of individuals in each popluation split
}
