% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.snpnet.R
\name{cv.snpnet}
\alias{cv.snpnet}
\title{Cross-validation for snpnet}
\usage{
cv.snpnet(
  genotype.pfile,
  phenotype.file,
  phenotype,
  family = NULL,
  covariates = NULL,
  weights = NULL,
  alpha = 1,
  nlambda = 100,
  lambda.min.ratio = NULL,
  full.lams = NULL,
  foldids = NULL,
  nfolds = 5,
  p.factor = NULL,
  ncores = 1,
  parallel = FALSE,
  status.col = NULL,
  mem = NULL,
  configs = NULL
)
}
\description{
Does k-fold cross-validation for snpnet, and returns a value for lambda
}
