% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_CI}
\alias{eval_CI}
\title{Compute R-squared or AUC with confidence interval.}
\usage{
eval_CI(data, response, predictors, family, level = 0.95)
}
\arguments{
\item{data}{A data frame containing the response and predictor(s)}

\item{response}{The name of the response variable}

\item{predictors}{The predictor variable(s)}

\item{level}{The confidence level of the interval}
}
\value{
a data frame containing r-squared or AUC (eval), confidence interval (l_eval, u_eval), p-value along with the information of the specified response and predictors
}
\description{
Given a data frame containing a response variable and predictor(s),
compute R-squared or AUC with confidence interval and the p-value.
We currently supports gaussian and binomial models.
If multiple predictors are specified, we return the minimum p-value
}
\examples{
\dontrun{
eval_CI(data_df, 'INI50', c('covar_score'), 'gaussian')
eval_CI(data_df, 'HC326', c('covar_score'), 'binomial')
}

}
