% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{FID_IID_from_rownames}
\alias{FID_IID_from_rownames}
\title{Recover FID and IID from rownames of the data frame}
\usage{
FID_IID_from_rownames(pheno_df, sep = "_")
}
\arguments{
\item{pheno_df}{A phenotype data frame with rownames}

\item{sep}{A separater used to concatenate FID and IID}
}
\value{
an updated phenotype data drame with FID and IID columns
}
\description{
Given a data frame of the phenotype data, recover FID and IID from
rownames
}
\examples{
\dontrun{
FID_IID_from_rownames(pheno_df)
}

}
