% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{FID_IID_to_rownames}
\alias{FID_IID_to_rownames}
\title{Assign rownames of phenotype data frame based on FID and IID}
\usage{
FID_IID_to_rownames(pheno_df, sep = "_")
}
\arguments{
\item{pheno_df}{A phenotype data frame with columns named 'FID' and 'IID'}

\item{sep}{A separater to concatenate FID and IID}
}
\value{
an updated phenotype data drame with rownames
}
\description{
Given a data frame of the phenotype data, concatenate FID and IID
columns with a separater character (sep) and use it as a rowname
}
\examples{
\dontrun{
FID_IID_to_rownames(pheno_df)
}

}
