% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_glm}
\alias{fit_glm}
\title{A wrapper function for glm()}
\usage{
fit_glm(data_df, formula_str, family)
}
\arguments{
\item{data_df}{A data frame containing the dataset for glm analysis}

\item{formula_str}{A string representing a regression formula}

\item{family}{The GLM family}
}
\value{
a glm fit object
}
\description{
Given a data frame, a regression formula in string, and a model family,
call glm function.
}
\examples{
\dontrun{
fit_glm(data, 'response ~ 1 + x + y', 'binomial')
}

}
