% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_glm_across_splits}
\alias{fit_glm_across_splits}
\title{Fit a specified regression model for each split independently
and aggregate the results into one data frame}
\usage{
fit_glm_across_splits(
  pheno_df,
  splits,
  glm_formula_str,
  family,
  split_col = "split"
)
}
\description{
Fit a specified regression model for each split independently
and aggregate the results into one data frame
}
