% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ID_ALTs}
\alias{get_ID_ALTs}
\title{Get ID_ALT of genetic variants}
\usage{
get_ID_ALTs(ID_ALT_file)
}
\arguments{
\item{ID_ALT_file}{a path to table (flat text) file}
}
\value{
list of ID_ALT values
}
\description{
read a table (text) file that contains "ID" and "ALT" column
and return as a list containing "ID_ALT"
}
