% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{glm_fit_to_r2}
\alias{glm_fit_to_r2}
\title{Get r-squared value from glm fit object}
\usage{
glm_fit_to_r2(glm_fit)
}
\arguments{
\item{glm_fit}{A fit object from glm()}
}
\value{
The r-squared value
}
\description{
Get r-squared value from glm fit object
}
