% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_covariates}
\alias{parse_covariates}
\title{Parse a string containing the set of covariates.}
\usage{
parse_covariates(covariates_str)
}
\arguments{
\item{covariates_str}{a string containing a list of covariates.}
}
\value{
List of covariates.
}
\description{
We assume the covariate_str contains set of covariates separated with ','.
}
