% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_PRS_binomial}
\alias{plot_PRS_binomial}
\title{generate Violin plot comparing the distribution of PRSs (Z-score if geno_z == TRUE)
stratified by case/control status.
plot_df should contain PRS and phenotype columns for (typically test) set of individuals
the column names can be specified in geno_score_col and phe_col}
\usage{
plot_PRS_binomial(
  plot_df,
  geno_score_col = "geno_score",
  phe_col = "phe",
  geno_z = TRUE
)
}
\description{
generate Violin plot comparing the distribution of PRSs (Z-score if geno_z == TRUE)
stratified by case/control status.
plot_df should contain PRS and phenotype columns for (typically test) set of individuals
the column names can be specified in geno_score_col and phe_col
}
