% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_PRS_vs_phe}
\alias{plot_PRS_vs_phe}
\title{generate a 2d heatmap plot comparing the PRS (Z-score if geno_z == TRUE) vs phenotype.}
\usage{
plot_PRS_vs_phe(
  plot_df,
  plot_bin2d_x = NULL,
  plot_bin2d_y = NULL,
  geno_score_col = "geno_score",
  phe_col = "phe",
  geno_z = TRUE
)
}
\arguments{
\item{plot_df}{PRS and phenotype columns for (typically test) set of individuals}

\item{plot_bin2d_x}{size of the x-axis grid in 2d heatmap}

\item{plot_bin2d_y}{size of the y-axis grid in 2d heatmap}

\item{geno_score_col}{column name of the score in plot_df}

\item{phe_col}{column name of the phenotype data in plot_df}

\item{geno_z}{whether to take the Z-score transformation of the score}
}
\description{
generate a 2d heatmap plot comparing the PRS (Z-score if geno_z == TRUE) vs phenotype.
}
