% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_external}
\alias{predict_external}
\title{Predict from using external weights}
\usage{
predict_external(variant_coefs, new_genotype_file, new_phenotype_file,
  phenotype, intercept = NULL, covariate_coefs = NULL, split_col = NULL, split_name = NULL, family = NULL, configs = list(zstdcat.path = "zstdcat", zcat.path = 'zcat'))
}
\arguments{
\item{variant_coefs}{the named vector of variant coefficents or weights where the names must match what is the genotype file}

\item{new_genotype_file}{Path to the new suite of genotype files. new_genotype_file.{pgen, psam, pvar.zst} must exist.}

\item{new_phenotype_file}{path to the phenotype. The header must include FID, IID. Used for extracting covaraites and computing metrics.}

\item{phenotype}{name of the phenotype for which the fit was computed.}

\item{intercept}{the intercept to use, zero by default}

\item{covariate_coefs}{a named vector of covariate coefficents, zero by default}

\item{split_col}{name of the split column. If NULL, all samples will be used.}

\item{split_name}{vector of split labels where prediction is to be made. Should be a combination of "train" (default), "val", "test".}

\item{family}{Type of the phenotype: "gaussian" for continuous phenotype and "binomial" for binary phenotype.}

\item{status_col}{name of the status column for Cox model.}

\item{configs}{additional list of configs including path to either zstdcat or zcat.}
}
\value{
a list containing the prediction and the response for which the prediction is made.
}
\description{
Predict from using external weights
}
