% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read_config_from_file}
\alias{read_config_from_file}
\title{Read a configuration from a file}
\usage{
read_config_from_file(config_file)
}
\arguments{
\item{config_file}{a path to a config file}
}
\value{
configugarion stored in a named list
}
\description{
Parse the 2-column config file for snpnet wrapper script
and return as a named list.
}
\details{
We parse the covariate string using parse_covariates() function
and cast the data type based on the type definitions stored in
config_params_data_type() function.
}
