% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read_lambda_sequence_from_RData_file}
\alias{read_lambda_sequence_from_RData_file}
\title{Load a lambda sequence from a R Data file}
\usage{
read_lambda_sequence_from_RData_file(rdata_f)
}
\arguments{
\item{rdata_f}{a file path to a R Data file}
}
\value{
lambda sequence
}
\description{
Read a R data file (saved from a snpnet run with validation set)
and load the lambda sequence up to the optimal lambda index
This helper function is useful to for refit of the snpnet model
using the individuals in combined set of training and validation sets.
}
