% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read_phenotype_file}
\alias{read_phenotype_file}
\title{Read a phenotype file as a data frame}
\usage{
read_phenotype_file(phenotype_file_path, columns)
}
\arguments{
\item{phenotype_file_path}{A path to the phenotype file}

\item{columns}{a list of columns to read}
}
\value{
data frame containing phenotype data
}
\description{
Read a phenotype file as a data frame for the selected columns
along with 'FID', 'IID', 'population', and 'split' columns
}
\details{
We recode -9 as NA in population, split, age, and sex columns
}
\examples{
\dontrun{
read_phenotype_file(phenotype_file_path, 'INI50')
read_phenotype_file(phenotype_file_path, c('INI50', 'HC382'))
read_phenotype_file(phenotype_file_path, c('age', 'sex', 'INI50', 'HC382'))
}

}
