% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{read_sscore_file}
\alias{read_sscore_file}
\title{Read a PRS file as a data frame}
\usage{
read_sscore_file(sscore_file_path, columns = c("SCORE1_SUM"))
}
\arguments{
\item{sscore_file_path}{A path to the sscore file}

\item{columns}{a list of columns to read}
}
\value{
data frame containing phenotype data
}
\description{
Read a sscore file (output from plink2 --score function) as a data frame
for the specified columns along with 'FID' and 'IID'
}
\examples{
\dontrun{
read_sscore_file(sscore_file_path)
}

}
