% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readPheMaster}
\alias{readPheMaster}
\title{Read covariates and phenotype(s) from the provided file path}
\usage{
readPheMaster(
  phenotype.file,
  psam.ids,
  family,
  covariates,
  phenotype,
  status,
  split.col,
  configs
)
}
\description{
Read covariates and phenotype(s) from the provided file path. Exclude individuals that contain
any missing value in the covariates, miss all phenotype values or do not have corresponding
genotypes.
}
