% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{recode_pheno_values}
\alias{recode_pheno_values}
\title{Recode phenotype values}
\usage{
recode_pheno_values(pheno_df, phes_binary = NULL, phes_quantitative = NULL)
}
\arguments{
\item{pheno_df}{A phenotype data frame}

\item{phes_binary}{A list of binary phenotypes}

\item{phes_quantitative}{A list of quantitative phenotypes}
}
\value{
an updated phenotype data drame
}
\description{
Given a data frame of phenotypes, we apply the following operations:
 1) replace -9 with NA
 2) for binary phenotypes, we subtract 1 so that
    the phenotype values takes [1, 0] instead of [2, 1]
Note: you need to specify the list of binary and quantitative phenotypes
using the phes_binary and phes_quantitative args.
}
\examples{
\dontrun{
recode_pheno_values(pheno_df, phes_binary=c('HC326'), phes_quantitative=c('INI50')
}

}
