% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{save_BETA}
\alias{save_BETA}
\title{save the regression coefficients (betas) into file(s).}
\usage{
save_BETA(df, out.file.head, pvar, covariates = NULL, verbose = FALSE)
}
\arguments{
\item{df}{a dataframe containing beta values}

\item{out.file.head}{a prefix of the output file name}

\item{pvar}{a path to pvar file containing the list of variants}

\item{covariates}{a list of covariates}

\item{verbose}{A boolean variable indicating whether we should dump logs}
}
\description{
We use up to 2 files corresponding to the betas for genetic variants
as well as the ones for covariates (optional).
}
\details{
For the genetic variants, we recover the information
stored in the pvar[.zst] file and sort the results
based on CHROM and POS columns.
}
