% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{snpnet_fit_to_df}
\alias{snpnet_fit_to_df}
\title{convert the snpnet regression coeffieicnets (betas) into a data frame.}
\usage{
snpnet_fit_to_df(beta, lambda_idx, covariates = NULL, verbose = FALSE)
}
\arguments{
\item{beta}{A list of vevtors containing beta values
at different lambda indices}

\item{lambda_idx}{A lambda index for which we extract the beta values}

\item{covariates}{A list of covariates}

\item{verbose}{A boolean variable indicating whether we should dump logs}
}
\value{
a data frame containing betas
}
\description{
convert the snpnet regression coeffieicnets (betas) into a data frame.
}
