% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_list_str}
\alias{split_list_str}
\title{split a list string into a list}
\usage{
split_list_str(list_str, pattern = ",")
}
\arguments{
\item{list_str}{A string object}
}
\value{
list
}
\description{
This is a wrapper for stringr::str_split()
}
\examples{
split_list_str('a,b,c')
split_list_str('a;b;c', ';')

}
