% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_named_list_str}
\alias{split_named_list_str}
\title{split a list string into a named list}
\usage{
split_named_list_str(named_list_str, pattern1 = ",", pattern2 = "=")
}
\arguments{
\item{named_list_str}{A string object}
}
\value{
named list
}
\description{
This is a wrapper for stringr::str_split()
}
\examples{
lapply(split_named_list_str('a=1,b=2,c=3'), as.numeric)
lapply(split_named_list_str('x=1;y=2;z=3', pattern1=';'), as.numeric)
lapply(split_named_list_str('p-1;q-2;r-3', pattern1=';', pattern2 = '-'), as.numeric)

}
