% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_split_column_for_refit}
\alias{update_split_column_for_refit}
\title{Collapse the training (train) and validation (val) sets into 'train_val'}
\usage{
update_split_column_for_refit(pheno_df)
}
\arguments{
\item{pheno_df}{A phenotype data frame with a column named 'split'}
}
\value{
an updated phenotype data drame
}
\description{
Given a data frame of the phenotype data, look at the 'split' column,
and replace 'train' and 'val' with 'train_val'
}
\examples{
\dontrun{
update_split_column_for_refit(pheno_df)
}

}
