% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpnet.R
\name{snpnet}
\alias{snpnet}
\title{Fit the Lasso for Large Phenotype-Genotype Datasets}
\usage{
snpnet(genotype.dir, phenotype.file, phenotype, results.dir = NULL,
  niter = 10, family = NULL, standardize.variant = FALSE, nlambda = 100,
  validation = FALSE, covariates = c("age", "sex", paste0("PC", 1:10)),
  num.snps.batch = 1000, glmnet.thresh = 1e-07, configs, verbose = TRUE,
  save = FALSE, use.glmnetPlus = (family == "gaussian"),
  early.stopping = TRUE, stopping.lag = 2, KKT.verbose = F,
  prevIter = 0)
}
\arguments{
\item{genotype.dir}{the directory that contains genotype files. Assume at least the existence of
train.bed/bim/fam, and val.bed/bim/fam if validation option is on.}

\item{phenotype.file}{the path of the file that contains the phenotype values and can be read as
as a table.}

\item{phenotype}{the name of the phenotype. Must be the same as the corresponding column name in
the phenotype file.}

\item{results.dir}{the path to the directory where meta and intermediate results are saved.}

\item{niter}{the number of maximum iteration in the algorithm}

\item{family}{the type of the phenotype: "gaussian" or "binomial". If not provided or NULL, it will be
detected based on the number of levels in the response}

\item{standardize.variant}{a logical value indicating whether the variants are standardized in
the lasso fitting. Default is FALSE. For SNP matrix, we may not want to standardize since the
variants are already on the same scale.}

\item{nlambda}{the number of lambda values on the solution path. Default is 100.}

\item{validation}{a logical value indicating if performance is evaluated on the validation set. If so,
val.bed/bim/fam should be available in genotype.dir.}

\item{covariates}{a character vector containing the names of the covariates included in the lasso
fitting, whose coefficients will not be penalized. The names must exist in the
column names of the phenotype file.}

\item{num.snps.batch}{the number of variants added to the strong set in each iteration. Default is 1000.}

\item{glmnet.thresh}{the convergence threshold for glmnet/glmnetPlus}

\item{save}{a logical value whether to save intermediate results (e.g. in case of job failure and restart)}

\item{use.glmnetPlus}{a logical value whether to use glmnet with warm start, if the glmnetPlus
package is available. Currently only "gaussian" family is supported}

\item{early.stopping}{a logical value indicating whether early stopping based on validation
metric is desired}

\item{stopping.lag}{a parameter for the stopping criterion such that the procedure stops after
this number of consecutive decreases in the validation performance}

\item{KKT.verbose}{a logical value indicating if details on KKT check should be printed}

\item{prevIter}{if non-zero, it indicates the last successful iteration in the procedure so that
we can restart from there. niter should be no less than prevIter.}
}
\value{
a list containing the solution path, the metric evaluated on training/validation set.
}
\description{
Fit the entire lasso solution path in an iterative way on large genomics datasets, in particular
phenotype-genotype datasets.
}
